requirements.txt
kafka-python==2.0.2
psycopg2-binary==2.9.9
streamlit==1.30.0
pandas==2.2.1
